refactor: cellix-tdd alignment for server-oauth2-mock-seedwork
- Align manifest.md and README.md with current public contract
- Add logger.test.ts for debugLog coverage
- Tighten internal module boundaries per cellix-tdd review
- All 126 seedwork tests and 42 app tests passing
