+++ packages/cellix/server-oauth2-mock-seedwork/manifest.md
@@ -1,23 +1,59 @@
 # manifest.md - @cellix/server-oauth2-mock-seedwork
 
-Package: @cellix/server-oauth2-mock-seedwork
+## Purpose
 
-Purpose:
-- Lightweight seedwork providing a mock OAuth2/OIDC server for local development and tests.
+Provide reusable mock OAuth2/OIDC server seedwork for local development, integration tests, and browser login simulations.
 
-Public exports:
-- startMockOAuth2Server(config): Promise<MockOAuth2ServerHandle>
-- createMockOAuth2Manager(): manager capable of registering multiple named configs
-- buildOidcRouter(issuerBaseUrl, config): Express router for embedding the mock OIDC endpoints
-- debugLog(message, data?): shared debug logger gated by `MOCK_OAUTH2_DEBUG`
+## Scope
 
-Downstream consumers (within monorepo):
-- @apps/server-oauth2-mock (app-level runnable)
+This package owns the public router, server bootstrap helpers, multi-registration manager, URL helpers, debug logging helper, and public configuration/data types needed to host a mock OIDC issuer.
 
-Non-goals:
-- Production-ready IdP replacement
-- Persistent token storage
+## Non-goals
 
-Notes:
-- Addition of createMockOAuth2Manager is additive and preserves backward compatibility with startMockOAuth2Server.
-- Tests cover multi-registration and token isolation behaviors.
+- Production-ready identity provider behavior
+- Durable token storage or revocation infrastructure
+- User-management UI beyond the minimal local-development login/signup forms
+
+## Public API shape
+
+- `startMockOAuth2Server(config): Promise<MockOAuth2ServerHandle>`
+- `createMockOAuth2Manager({ port, host?, baseUrl }): MockOAuth2Manager`
+- `buildOidcRouter(issuerBaseUrl, config): Promise<express.Router>`
+- `debugLog(message, data?): void`
+- `normalizeBaseUrl(url)`, `normalizeOrigin(url)`, `normalizeUrl(url)`, `SAFE_NAME_RE`
+- Public types from `src/types.ts`, especially `MockOAuth2PortalConfig`, `MockOAuth2ServerConfig`, and the async `MockOAuth2UserStore`
+
+## Core concepts
+
+- One issuer can be hosted directly with `startMockOAuth2Server`, or many named issuers can share one process via `createMockOAuth2Manager`.
+- `MockOAuth2UserStore` is intentionally async so downstream stores can perform filesystem or database I/O on the request path.
+- Interactive login/signup flows persist short-lived nonces server-side and exchange them for auth codes without echoing untrusted query data into HTML.
+
+## Package boundaries
+
+- Keep login/signup route composition, HTML builders, TTL stores, and claim-merging helpers internal unless an external consumer has a clear need.
+- `src/login-handlers.ts` is internal implementation detail extracted only to keep `router.ts` focused.
+- The package should not expose app-specific file-store behavior or monorepo-only bootstrapping concerns.
+
+## Dependencies / relationships
+
+- Downstream consumer in this monorepo: `@apps/server-oauth2-mock`
+- Depends on `express`, `express-rate-limit`, and `jose` for HTTP routing, rate limiting, and JWT/JWKS behavior
+
+## Testing strategy
+
+- Prefer public-entrypoint tests that exercise `buildOidcRouter`, `createMockOAuth2Manager`, and `startMockOAuth2Server` end-to-end.
+- Preserve contract tests for interactive login/signup, token issuance, userinfo, multi-registration isolation, and URL helper behavior.
+- Avoid adding tests for internal helpers when an existing public-contract test already proves the behavior.
+
+## Documentation obligations
+
+- Keep `README.md` consumer-facing and package-centric.
+- Keep TSDoc aligned for meaningful public exports and public types.
+- Update this manifest whenever package boundaries, public exports, or release caveats change.
+
+## Release-readiness standards
+
+- Public entrypoint exports remain intentionally narrow.
+- Package build and existing test suites for the seedwork and affected downstream consumer must pass.
+- Any breaking change to `MockOAuth2UserStore` or other public exports requires explicit semver review before release.
