+++ packages/cellix/server-oauth2-mock-seedwork/src/logger.test.ts
@@ -1,0 +1,8 @@
+import { describe, expect, it } from 'vitest';
+import { debugLog } from './index.ts';
+
+describe('debugLog', () => {
+	it('returns undefined when debug logging is disabled', () => {
+		expect(debugLog('test message')).toBeUndefined();
+	});
+});
