+++ packages/cellix/server-oauth2-mock-seedwork/src/utils.ts
@@ -1,7 +1,31 @@
-/** Utility helpers for server-oauth2-mock-seedwork */
+/** Utility helpers for server-oauth2-mock-seedwork. */
 
+/**
+ * Regular expression used to validate named portal registrations.
+ *
+ * @returns A regular expression that matches only letters, digits, underscores, and hyphens.
+ *
+ * @example
+ * ```ts
+ * SAFE_NAME_RE.test('portal_admin-1'); // true
+ * SAFE_NAME_RE.test('portal/admin'); // false
+ * ```
+ */
 export const SAFE_NAME_RE = /^[a-zA-Z0-9_-]+$/;
 
+/**
+ * Normalizes a URL for redirect-URI comparisons by removing a trailing slash from the
+ * path and sorting query parameters.
+ *
+ * @param value - The candidate URL string.
+ * @returns The normalized URL string, or the original value when parsing fails.
+ *
+ * @example
+ * ```ts
+ * normalizeUrl('https://example.com/callback/?b=2&a=1');
+ * // => 'https://example.com/callback?a=1&b=2'
+ * ```
+ */
 export const normalizeUrl = (value: string) => {
 	try {
 		const url = new URL(value);
@@ -15,6 +39,18 @@ export const normalizeUrl = (value: string) => {
 	}
 };
 
+/**
+ * Extracts the origin portion of a URL so CORS checks can compare normalized origins.
+ *
+ * @param value - The candidate URL string.
+ * @returns The normalized origin, or the original value when parsing fails.
+ *
+ * @example
+ * ```ts
+ * normalizeOrigin('https://example.com/path?x=1');
+ * // => 'https://example.com'
+ * ```
+ */
 export const normalizeOrigin = (value: string) => {
 	try {
 		const url = new URL(value);
@@ -24,6 +60,18 @@ export const normalizeOrigin = (value: string) => {
 	}
 };
 
+/**
+ * Removes a trailing slash from a configured issuer or base URL.
+ *
+ * @param url - The configured base URL.
+ * @returns The base URL without a trailing slash.
+ *
+ * @example
+ * ```ts
+ * normalizeBaseUrl('http://localhost:3000/');
+ * // => 'http://localhost:3000'
+ * ```
+ */
 export function normalizeBaseUrl(url: string): string {
 	return url.replace(/\/$/, '');
 }
