+++ packages/cellix/server-oauth2-mock-seedwork/tests/invalid-name.test.ts
@@ -10,30 +10,32 @@ function makeConfig(): MockOAuth2PortalConfig {
 	};
 }
 
-describe('portal name validation', () => {
-	it('rejects unsafe names with path segments', async () => {
-		const manager = createMockOAuth2Manager({ port: 19020, host: '127.0.0.1', baseUrl: 'http://127.0.0.1:19020' });
-		const cfg = makeConfig();
-		await expect(manager.register('foo/bar', cfg)).rejects.toThrow(`[server-oauth2-mock] Invalid portal name "foo/bar": must contain letters, digits, '_' and '-' only`);
-	});
+describe('createMockOAuth2Manager', () => {
+	describe('portal name validation', () => {
+		it('rejects unsafe names with path segments', async () => {
+			const manager = createMockOAuth2Manager({ port: 19020, host: '127.0.0.1', baseUrl: 'http://127.0.0.1:19020' });
+			const cfg = makeConfig();
+			await expect(manager.register('foo/bar', cfg)).rejects.toThrow(`[server-oauth2-mock] Invalid portal name "foo/bar": must contain letters, digits, '_' and '-' only`);
+		});
 
-	it('rejects unsafe names with traversal segments', async () => {
-		const manager = createMockOAuth2Manager({ port: 19021, host: '127.0.0.1', baseUrl: 'http://127.0.0.1:19021' });
-		const cfg = makeConfig();
-		await expect(manager.register('../admin', cfg)).rejects.toThrow(`[server-oauth2-mock] Invalid portal name "../admin": must contain letters, digits, '_' and '-' only`);
-	});
+		it('rejects unsafe names with traversal segments', async () => {
+			const manager = createMockOAuth2Manager({ port: 19021, host: '127.0.0.1', baseUrl: 'http://127.0.0.1:19021' });
+			const cfg = makeConfig();
+			await expect(manager.register('../admin', cfg)).rejects.toThrow(`[server-oauth2-mock] Invalid portal name "../admin": must contain letters, digits, '_' and '-' only`);
+		});
 
-	it('rejects names with disallowed characters such as dots', async () => {
-		const manager = createMockOAuth2Manager({ port: 19022, host: '127.0.0.1', baseUrl: 'http://127.0.0.1:19022' });
-		const cfg = makeConfig();
-		await expect(manager.register('foo.bar', cfg)).rejects.toThrow(`[server-oauth2-mock] Invalid portal name "foo.bar": must contain letters, digits, '_' and '-' only`);
-	});
+		it('rejects names with disallowed characters such as dots', async () => {
+			const manager = createMockOAuth2Manager({ port: 19022, host: '127.0.0.1', baseUrl: 'http://127.0.0.1:19022' });
+			const cfg = makeConfig();
+			await expect(manager.register('foo.bar', cfg)).rejects.toThrow(`[server-oauth2-mock] Invalid portal name "foo.bar": must contain letters, digits, '_' and '-' only`);
+		});
 
-	it('accepts names with letters, digits, underscores, and hyphens', async () => {
-		const manager = createMockOAuth2Manager({ port: 19023, host: '127.0.0.1', baseUrl: 'http://127.0.0.1:19023' });
-		const cfg = makeConfig();
-		const result = await manager.register('admin-portal_1', cfg);
-		expect(result.name).toBe('admin-portal_1');
-		await manager.stopAll();
+		it('accepts names with letters, digits, underscores, and hyphens', async () => {
+			const manager = createMockOAuth2Manager({ port: 19023, host: '127.0.0.1', baseUrl: 'http://127.0.0.1:19023' });
+			const cfg = makeConfig();
+			const result = await manager.register('admin-portal_1', cfg);
+			expect(result.name).toBe('admin-portal_1');
+			await manager.stopAll();
+		});
 	});
 });
